+++ src/Paysera/WalletApi/Client/TokenRelatedWalletClient.php
@@ -51,7 +51,7 @@ class Paysera_WalletApi_Client_TokenRelatedWalletClient extends Paysera_WalletAp
      *
      * @return Paysera_WalletApi_Entity_Search_Result|Paysera_WalletApi_Entity_Statement[]
      */
-    public function getCurrentWalletStatements(Paysera_WalletApi_Entity_Statement_SearchFilter $filter = null)
+    public function getCurrentWalletStatements(?Paysera_WalletApi_Entity_Statement_SearchFilter $filter = null)
     {
         return $this->getWalletStatements('me', $filter);
     }