+++ src/Paysera/WalletApi/Listener/ParameterSetter.php
@@ -32,7 +32,7 @@ class Paysera_WalletApi_Listener_ParameterSetter implements Paysera_WalletApi_Ev
         $event->setOptions($options);
     }
 
-    public static function getSubscribedEvents()
+    public static function getSubscribedEvents(): array
     {
         return array(
             Paysera_WalletApi_Events::BEFORE_REQUEST => array('onBeforeRequest', 100),