+++ tests/Paysera/WalletApi/Service/Paysera_WalletApi_Entity_LocationTest.php
@@ -4,14 +4,14 @@
  * Date: 2014-02-25
  */
 
-class Paysera_WalletApi_Service_LocationManagerTest extends PHPUnit_Framework_TestCase
+class Paysera_WalletApi_Service_LocationManagerTest extends PHPUnit\Framework\TestCase
 {
     /**
      * @var Paysera_WalletApi_Service_LocationManager
      */
     protected $service;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->service = new Paysera_WalletApi_Service_LocationManager();
     }